+++ server/db.json
@@ -1,0 +1,1 @@
+Datos de prueba
