+++ src/router/AbisalRoutes.jsx
@@ -1,0 +1,1 @@
+import { createBrowserRouter } from "react-router-dom";
